Creates an associations table.

Creates an associations table.

Usage

associations_tbl(
  association_id = character(),
  pvalue = double(),
  pvalue_description = character(),
  pvalue_mantissa = integer(),
  pvalue_exponent = integer(),
  multiple_snp_haplotype = logical(),
  snp_interaction = logical(),
  snp_type = character(),
  risk_frequency = double(),
  standard_error = double(),
  range = character(),
  or_per_copy_number = double(),
  beta_number = double(),
  beta_unit = character(),
  beta_direction = character(),
  beta_description = character(),
  last_mapping_date = lubridate::ymd_hms(),
  last_update_date = lubridate::ymd_hms()
)

Arguments

association_id

A character vector of association identifiers.

pvalue

A numeric vector of p-values.

pvalue_description

A character vector of p-value context descriptions.

pvalue_mantissa

An integer vector of p-value mantissas.

pvalue_exponent

An integer vector of p-value exponents.

multiple_snp_haplotype

A logical vector.

snp_interaction

A logical vector.

snp_type

A character vector indicating SNP novelty: 'novel' or 'known'.

standard_error

A numeric vector of standard errors.

range

A character vector of free text descriptions of confidence intervals.

or_per_copy_number

A numeric vector of odds ratios.

beta_number

A numeric vector of beta coefficients.

beta_unit

A character vector of beta coefficient units.

beta_direction

A character vector of beta coefficient directions.

beta_description

A character vector of beta descriptions.

last_mapping_date

A POSIXct object indicating last time this association was mapped to Ensembl.

last_update_date

A POSIXct object indicating the last time the associations have been updated.

Value

A tibble whose columns are the named arguments to the function.