Get GWAS Catalog metadata

Provides a list of the resources the GWAS Catalog data is currently mapped against: Ensembl release number, Genome build version and dbSNP version.In addition, the date since this combination of resource versions has been in use is also returned.

get_metadata(verbose = FALSE, warnings = TRUE)

Arguments

verbose
Whether to be chatty.

warnings
Whether to trigger a warning if the request is not successful.

Value

A named list whose names are:

ensembl_release_number: Ensembl release number;

genome_build_version: Genome build version;

dbsnp_version: dbSNP version.

usage_start_date: Date since this combination of resource versions has been in use.

Examples

get_metadata()

#> $ensembl_release_number
#> [1] 101
#> 
#> $genome_build_version
#> [1] "GRCh38.p13"
#> 
#> $dbsnp_version
#> [1] 153
#> 
#> $usage_start_date
#> [1] "2020-09-11 13:00:00 UTC"
#>