Get GWAS Catalog variants

Retrieves variants via the NHGRI-EBI GWAS Catalog REST API. The REST API is queried multiple times with the criteria passed as arguments (see below). By default all variants that match the criteria supplied in the arguments are retrieved: this corresponds to the default option set_operation set to 'union'. If you rather have only the variants that match simultaneously all criteria provided, then set set_operation to 'intersection'.

get_variants(
  study_id = NULL,
  association_id = NULL,
  variant_id = NULL,
  efo_id = NULL,
  pubmed_id = NULL,
  genomic_range = NULL,
  cytogenetic_band = NULL,
  gene_name = NULL,
  efo_trait = NULL,
  reported_trait = NULL,
  set_operation = "union",
  interactive = TRUE,
  std_chromosomes_only = TRUE,
  verbose = FALSE,
  warnings = TRUE
)

Arguments

study_id
A character vector of GWAS Catalog study accession identifiers.

association_id
A character vector of GWAS Catalog association identifiers.

variant_id
A character vector of GWAS Catalog variant identifiers.

efo_id
A character vector of EFO identifiers.

pubmed_id
An integer vector of PubMed identifiers.

genomic_range
A named list of three vectors:

chromosome

A character vector of chromosome names of the form 1--22, X or Y.

start

A numeric vector of start positions, starting at 1.

end

A numeric vector of end positions.

The three vectors need to be of the same length so that chromosome names, start and end positions can be matched by position.

cytogenetic_band
A character vector of cytogenetic bands of the form '1p36.11'.

gene_name
Gene symbol according to HUGO Gene Nomenclature (HGNC).

efo_trait
A character vector of EFO trait descriptions, e.g., 'uric acid measurement'.

reported_trait
A character vector of phenotypic traits as reported by the original authors of the study.

set_operation
Either 'union' or 'intersection'. This tells how variants retrieved by different criteria should be combined: 'union' binds together all results removing duplicates and 'intersection' only keeps same variants found with different criteria.

interactive
A logical. If all variants are requested, whether to ask interactively if we really want to proceed.

std_chromosomes_only
Whether to return only variants mapped to standard chromosomes: 1 thru 22, X, Y, and MT.

verbose
Whether the function should be verbose about the different queries or not.

warnings
Whether to print warnings.

Value

A variants object.

Details

Please note that all search criteria are vectorised, thus allowing for batch mode search, e.g., one can search by multiple variant identifiers at once by passing a vector of identifiers to variant_id.

Examples

# Get variants by study identifier
get_variants(study_id = 'GCST001085')
#> An object of class "variants"
#> Slot "variants":
#> # A tibble: 18 × 7
#>    variant_id merged functional_class          chromosome_name chromosome_posit…
#>    <chr>       <int> <chr>                     <chr>                       <int>
#>  1 rs10496288      0 intergenic_variant        NA                             NA
#>  2 rs7827545       0 intron_variant            NA                             NA
#>  3 rs10785581      0 intron_variant            12                       45537190
#>  4 rs200759        0 intron_variant            20                       15625776
#>  5 rs7960483       0 intron_variant            NA                             NA
#>  6 rs1372662       0 intron_variant            8                       134554803
#>  7 rs10496289      0 intergenic_variant        2                        83066256
#>  8 rs6452524       0 intron_variant            NA                             NA
#>  9 rs200752        0 intron_variant            NA                             NA
#> 10 rs3798440       0 intron_variant            NA                             NA
#> 11 rs7735940       0 regulatory_region_variant NA                             NA
#> 12 rs9350602       0 intron_variant            6                        75850781
#> 13 rs10188442      0 intron_variant            2                       132431666
#> 14 rs12522034      0 intergenic_variant        5                        36425491
#> 15 rs13420028      0 intron_variant            NA                             NA
#> 16 rs2469997       0 intergenic_variant        8                       119341027
#> 17 rs6887846       0 intron_variant            5                        83140542
#> 18 rs6469823       1 intergenic_variant        8                       119341744
#> # … with 2 more variables: chromosome_region <chr>, last_update_date <dttm>
#> 
#> Slot "genomic_contexts":
#> # A tibble: 60 × 12
#>    variant_id gene_name chromosome_name chromosome_posi… distance is_mapped_gene
#>    <chr>      <chr>     <chr>                      <int>    <int> <lgl>         
#>  1 rs10785581 LOC10536… 12                      45537190    58636 FALSE         
#>  2 rs10785581 ANO6      12                      45537190    96786 FALSE         
#>  3 rs10785581 ANO6      12                      45537190    54910 TRUE          
#>  4 rs10785581 U6        12                      45537190    17400 TRUE          
#>  5 rs10785581 LOC10012… 12                      45537190    91755 FALSE         
#>  6 rs10785581 LOC10536… 12                      45537190        0 FALSE         
#>  7 rs200759   MACROD2   20                      15625776        0 FALSE         
#>  8 rs200759   LOC10798… 20                      15625776     1048 FALSE         
#>  9 rs200759   ENSAP1    20                      15625776     5991 FALSE         
#> 10 rs200759   ENSAP1    20                      15625776     5957 FALSE         
#> # … with 50 more rows, and 6 more variables: is_closest_gene <lgl>,
#> #   is_intergenic <lgl>, is_upstream <lgl>, is_downstream <lgl>, source <chr>,
#> #   mapping_method <chr>
#> 
#> Slot "ensembl_ids":
#> # A tibble: 34 × 3
#>    variant_id gene_name ensembl_id     
#>    <chr>      <chr>     <chr>          
#>  1 rs10785581 ANO6      ENSG00000177119
#>  2 rs10785581 U6        ENSG00000264557
#>  3 rs10785581 U6        ENSG00000265920
#>  4 rs10785581 U6        ENSG00000263869
#>  5 rs10785581 U6        ENSG00000265351
#>  6 rs10785581 U6        ENSG00000266660
#>  7 rs10785581 U6        ENSG00000265048
#>  8 rs10785581 U6        ENSG00000266435
#>  9 rs200759   MACROD2   ENSG00000172264
#> 10 rs200759   ENSAP1    ENSG00000224274
#> # … with 24 more rows
#> 
#> Slot "entrez_ids":
#> # A tibble: 32 × 3
#>    variant_id gene_name    entrez_id
#>    <chr>      <chr>        <chr>    
#>  1 rs10785581 LOC105369744 105369744
#>  2 rs10785581 ANO6         196527   
#>  3 rs10785581 LOC100128248 100128248
#>  4 rs10785581 LOC105369743 105369743
#>  5 rs200759   MACROD2      140733   
#>  6 rs200759   LOC107985394 107985394
#>  7 rs200759   ENSAP1       170511   
#>  8 rs1372662  ZFAT-AS1     594840   
#>  9 rs1372662  ZFAT         57623    
#> 10 rs9350602  MYO6         4646     
#> # … with 22 more rows
#> 

# Get variants by association identifier
get_variants(association_id = '25389945')
#> An object of class "variants"
#> Slot "variants":
#> # A tibble: 27 × 7
#>    variant_id merged functional_class          chromosome_name chromosome_posit…
#>    <chr>       <int> <chr>                     <chr>                       <int>
#>  1 rs9486815       0 intron_variant            6                       108126842
#>  2 rs4245535       0 intron_variant            NA                             NA
#>  3 rs17069173      0 intergenic_variant        NA                             NA
#>  4 rs9374021       0 intron_variant            NA                             NA
#>  5 rs9374007       0 intergenic_variant        NA                             NA
#>  6 rs9386694       0 intergenic_variant        NA                             NA
#>  7 rs9374002       0 non_coding_transcript_ex… NA                             NA
#>  8 rs218289        0 intron_variant            NA                             NA
#>  9 rs1064346       0 3_prime_UTR_variant       NA                             NA
#> 10 rs9374013       0 intron_variant            NA                             NA
#> # … with 17 more rows, and 2 more variables: chromosome_region <chr>,
#> #   last_update_date <dttm>
#> 
#> Slot "genomic_contexts":
#> # A tibble: 10 × 12
#>    variant_id gene_name chromosome_name chromosome_posi… distance is_mapped_gene
#>    <chr>      <chr>     <chr>                      <int>    <int> <lgl>         
#>  1 rs9486815  SNX3      6                      108126842    84380 FALSE         
#>  2 rs9486815  SNX3      6                      108126842    84380 FALSE         
#>  3 rs9486815  NR2E1     6                      108126842    39180 FALSE         
#>  4 rs9486815  Y_RNA     6                      108126842    65912 FALSE         
#>  5 rs9486815  OSTM1     6                      108126842        0 TRUE          
#>  6 rs9486815  NR2E1     6                      108126842    39180 FALSE         
#>  7 rs9486815  OSTM1-AS1 6                      108126842        0 FALSE         
#>  8 rs9486815  LOC10537… 6                      108126842    86455 FALSE         
#>  9 rs9486815  OSTM1     6                      108126842    52101 FALSE         
#> 10 rs9486815  OSTM1-AS1 6                      108126842        0 TRUE          
#> # … with 6 more variables: is_closest_gene <lgl>, is_intergenic <lgl>,
#> #   is_upstream <lgl>, is_downstream <lgl>, source <chr>, mapping_method <chr>
#> 
#> Slot "ensembl_ids":
#> # A tibble: 5 × 3
#>   variant_id gene_name ensembl_id     
#>   <chr>      <chr>     <chr>          
#> 1 rs9486815  SNX3      ENSG00000112335
#> 2 rs9486815  NR2E1     ENSG00000112333
#> 3 rs9486815  Y_RNA     ENSG00000202533
#> 4 rs9486815  OSTM1     ENSG00000081087
#> 5 rs9486815  OSTM1-AS1 ENSG00000225174
#> 
#> Slot "entrez_ids":
#> # A tibble: 5 × 3
#>   variant_id gene_name    entrez_id
#>   <chr>      <chr>        <chr>    
#> 1 rs9486815  SNX3         8724     
#> 2 rs9486815  NR2E1        7101     
#> 3 rs9486815  OSTM1        28962    
#> 4 rs9486815  OSTM1-AS1    100287366
#> 5 rs9486815  LOC105377929 105377929
#> 

# Get a variant by its identifier
get_variants(variant_id = 'rs3798440')
#> An object of class "variants"
#> Slot "variants":
#> # A tibble: 1 × 7
#>   variant_id merged functional_class chromosome_name chromosome_position
#>   <chr>       <int> <chr>            <chr>                         <int>
#> 1 rs3798440       0 intron_variant   NA                               NA
#> # … with 2 more variables: chromosome_region <chr>, last_update_date <dttm>
#> 
#> Slot "genomic_contexts":
#> # A tibble: 0 × 12
#> # … with 12 variables: variant_id <chr>, gene_name <chr>,
#> #   chromosome_name <chr>, chromosome_position <int>, distance <int>,
#> #   is_mapped_gene <lgl>, is_closest_gene <lgl>, is_intergenic <lgl>,
#> #   is_upstream <lgl>, is_downstream <lgl>, source <chr>, mapping_method <chr>
#> 
#> Slot "ensembl_ids":
#> # A tibble: 0 × 3
#> # … with 3 variables: variant_id <chr>, gene_name <chr>, ensembl_id <chr>
#> 
#> Slot "entrez_ids":
#> # A tibble: 0 × 3
#> # … with 3 variables: variant_id <chr>, gene_name <chr>, entrez_id <chr>
#> 

# Get variants by EFO trait identifier
get_variants(efo_id = 'EFO_0005537')
#> An object of class "variants"
#> Slot "variants":
#> # A tibble: 8 × 7
#>   variant_id merged functional_class    chromosome_name chromosome_position
#>   <chr>       <int> <chr>               <chr>                         <int>
#> 1 rs2464195       0 missense_variant    12                        120997672
#> 2 rs17215231      0 5_prime_UTR_variant 6                          33272092
#> 3 rs10069690      0 intron_variant      5                           1279675
#> 4 rs2363956       0 missense_variant    19                         17283315
#> 5 rs4245739       0 3_prime_UTR_variant 1                         204549714
#> 6 rs10771399      0 intergenic_variant  12                         28002147
#> 7 rs3757318       0 intron_variant      6                         151592978
#> 8 rs78378222      0 3_prime_UTR_variant 17                          7668434
#> # … with 2 more variables: chromosome_region <chr>, last_update_date <dttm>
#> 
#> Slot "genomic_contexts":
#> # A tibble: 359 × 12
#>    variant_id gene_name chromosome_name chromosome_posi… distance is_mapped_gene
#>    <chr>      <chr>     <chr>                      <int>    <int> <lgl>         
#>  1 rs2464195  OASL2P    12                     120997672    56060 FALSE         
#>  2 rs2464195  CLIC1P1   12                     120997672    82549 FALSE         
#>  3 rs2464195  OASL      12                     120997672    21210 FALSE         
#>  4 rs2464195  XLOC_009… 12                     120997672    89598 FALSE         
#>  5 rs2464195  HNF1A     12                     120997672        0 TRUE          
#>  6 rs2464195  RPL12P33  12                     120997672    80436 FALSE         
#>  7 rs2464195  LOC10537… 12                     120997672    71187 FALSE         
#>  8 rs2464195  C12orf43  12                     120997672     2814 FALSE         
#>  9 rs2464195  RPL12P33  12                     120997672    80387 FALSE         
#> 10 rs2464195  C12orf43  12                     120997672     2814 FALSE         
#> # … with 349 more rows, and 6 more variables: is_closest_gene <lgl>,
#> #   is_intergenic <lgl>, is_upstream <lgl>, is_downstream <lgl>, source <chr>,
#> #   mapping_method <chr>
#> 
#> Slot "ensembl_ids":
#> # A tibble: 154 × 3
#>    variant_id gene_name ensembl_id     
#>    <chr>      <chr>     <chr>          
#>  1 rs2464195  OASL2P    ENSG00000283542
#>  2 rs2464195  CLIC1P1   ENSG00000231313
#>  3 rs2464195  OASL      ENSG00000135114
#>  4 rs2464195  HNF1A     ENSG00000135100
#>  5 rs2464195  RPL12P33  ENSG00000244585
#>  6 rs2464195  C12orf43  ENSG00000157895
#>  7 rs2464195  SPPL3     ENSG00000157837
#>  8 rs2464195  HNF1A-AS1 ENSG00000241388
#>  9 rs17215231 ZBTB22    ENSG00000236104
#> 10 rs17215231 ZBTB22    ENSG00000206280
#> # … with 144 more rows
#> 
#> Slot "entrez_ids":
#> # A tibble: 89 × 3
#>    variant_id gene_name    entrez_id
#>    <chr>      <chr>        <chr>    
#>  1 rs2464195  OASL2P       111216278
#>  2 rs2464195  CLIC1P1      390363   
#>  3 rs2464195  OASL         8638     
#>  4 rs2464195  XLOC_009911  105500240
#>  5 rs2464195  HNF1A        6927     
#>  6 rs2464195  RPL12P33     643550   
#>  7 rs2464195  LOC105378258 105378258
#>  8 rs2464195  C12orf43     64897    
#>  9 rs2464195  SPPL3        121665   
#> 10 rs2464195  HNF1A-AS1    283460   
#> # … with 79 more rows
#>