Get GWAS Catalog variants by cytogenetic band.

Gets variants that are mapped onto specific regions as specified by cytogenetic bands. See the dataframe cytogenetic_bands for more information on possible values.

get_variants_by_cytogenetic_band(
  cytogenetic_band = NULL,
  verbose = FALSE,
  warnings = TRUE,
  page_size = 20L
)

Arguments

cytogenetic_band
A character vector of cytogenetic bands of the form '1p36.11'

verbose
Whether the function should be verbose about the different queries or not.

warnings
Whether to print warnings.

page_size
An integer scalar indicating the page value to be used in the JSON requests, can be between 1 and 1000.

Value

A variants object.