GWAS accession identifier mapping

Use these functions to map between identifiers of studies, associations, variants or traits. Each association_id maps to just one study_id; all other mappings are one-to-many. Moreover, in the case of the mapping of variant identifiers to trait identifiers (efo_id) the mapping is first performed from variant id to association id, and only then to trait identifier. All other mappings are direct in the sense that the Gwas Catalog REST API provides the endpoints that make those direct queries possible.

study_to_association(study_id, verbose = FALSE, warnings = TRUE)

study_to_variant(study_id, verbose = FALSE, warnings = TRUE)

study_to_trait(study_id, verbose = FALSE, warnings = TRUE)

association_to_study(association_id, verbose = FALSE, warnings = TRUE)

association_to_variant(association_id, verbose = FALSE, warnings = TRUE)

association_to_trait(association_id, verbose = FALSE, warnings = TRUE)

variant_to_study(variant_id, verbose = FALSE, warnings = TRUE)

variant_to_association(variant_id, verbose = FALSE, warnings = TRUE)

variant_to_trait(
  variant_id,
  keep_association_id = FALSE,
  verbose = FALSE,
  warnings = TRUE
)

trait_to_study(efo_id, verbose = FALSE, warnings = TRUE)

trait_to_association(efo_id, verbose = FALSE, warnings = TRUE)

trait_to_variant(efo_id, verbose = FALSE, warnings = TRUE)

Arguments

study_id, association_id, variant_id, efo_id
A character vector of accession identifiers for either studies (study_id), associations (association_id), variants (variant_id) or traits (efo_id).

verbose
Whether the function should be verbose about the different queries or not.

warnings
Whether to print warnings.

keep_association_id
In the case of the variant_to_trait() function, the mapping from variant ids to EFO traits ids is performed via association ids; this option allow you to keep the association identifier in the final output (default is FALSE).

Value

A dataframe of two identifiers (three if keep_association_id is TRUE and if variant_to_trait() used). First column is the from identifier and the second column is the to identifier.

Examples

# Map GWAS study identifiers to association identifiers
study_to_association(c('GCST001084', 'GCST001085'))

#> # A tibble: 14 x 2
#>    study_id   association_id
#>    <chr>      <chr>         
#>  1 GCST001084 24300089      
#>  2 GCST001084 24300097      
#>  3 GCST001084 24300105      
#>  4 GCST001084 24300113      
#>  5 GCST001084 24300121      
#>  6 GCST001085 24299694      
#>  7 GCST001085 24299702      
#>  8 GCST001085 24299710      
#>  9 GCST001085 24299718      
#> 10 GCST001085 24299726      
#> 11 GCST001085 24299734      
#> 12 GCST001085 24299743      
#> 13 GCST001085 24299751      
#> 14 GCST001085 24299759      

# Map GWAS study identifiers to variant identifiers
study_to_variant(c('GCST001084', 'GCST001085'))

#> # A tibble: 28 x 2
#>    study_id   variant_id
#>    <chr>      <chr>     
#>  1 GCST001084 rs7697839 
#>  2 GCST001084 rs1165668 
#>  3 GCST001084 rs1165669 
#>  4 GCST001084 rs10218795
#>  5 GCST001084 rs1333048 
#>  6 GCST001084 rs1333049 
#>  7 GCST001084 rs6789378 
#>  8 GCST001084 rs12091564
#>  9 GCST001084 rs7673097 
#> 10 GCST001084 rs11924705
#> # … with 18 more rows

# Map GWAS study identifiers to EFO trait identifiers
study_to_trait(c('GCST001084', 'GCST001085'))

#> # A tibble: 2 x 2
#>   study_id   efo_id     
#>   <chr>      <chr>      
#> 1 GCST001084 EFO_0001645
#> 2 GCST001085 EFO_0000537

# Map GWAS association identifiers to study identifiers
association_to_study(c('24300097', '24299759'))

#> # A tibble: 2 x 2
#>   association_id study_id  
#>   <chr>          <chr>     
#> 1 24300097       GCST001084
#> 2 24299759       GCST001085

# Map GWAS association identifiers to variant identifiers
association_to_variant(c('24300097', '24299759'))

#> # A tibble: 4 x 2
#>   association_id variant_id
#>   <chr>          <chr>     
#> 1 24300097       rs11924705
#> 2 24300097       rs6789378 
#> 3 24299759       rs10496288
#> 4 24299759       rs10496289

# Map GWAS association identifiers to EFO trait identifiers
association_to_trait(c('24300097', '24299759'))

#> # A tibble: 2 x 2
#>   association_id efo_id     
#>   <chr>          <chr>      
#> 1 24300097       EFO_0001645
#> 2 24299759       EFO_0000537

# Map GWAS variant identifiers to study identifiers
variant_to_study(c('rs7904579', 'rs138331350'))

#> # A tibble: 2 x 2
#>   variant_id  study_id  
#>   <chr>       <chr>     
#> 1 rs7904579   GCST005387
#> 2 rs138331350 GCST007140

# Map GWAS variant identifiers to association identifiers
variant_to_association(c('rs7904579', 'rs138331350'))

#> # A tibble: 3 x 2
#>   variant_id  association_id
#>   <chr>       <chr>         
#> 1 rs7904579   27460811      
#> 2 rs138331350 35840603      
#> 3 rs138331350 35840598      

# Map GWAS variant identifiers to EFO trait identifiers
variant_to_trait(c('rs7904579', 'rs138331350'))

#> # A tibble: 4 x 2
#>   variant_id  efo_id     
#>   <chr>       <chr>      
#> 1 rs7904579   EFO_0005924
#> 2 rs7904579   EFO_0000289
#> 3 rs138331350 EFO_0004612
#> 4 rs138331350 EFO_0004612

# Map GWAS variant identifiers to EFO trait identifiers
# but keep the intermediate association identifier
variant_to_trait(c('rs7904579', 'rs138331350'), keep_association_id = TRUE)

#> # A tibble: 4 x 3
#>   variant_id  association_id efo_id     
#>   <chr>       <chr>          <chr>      
#> 1 rs7904579   27460811       EFO_0005924
#> 2 rs7904579   27460811       EFO_0000289
#> 3 rs138331350 35840603       EFO_0004612
#> 4 rs138331350 35840598       EFO_0004612

# Map EFO trait identifiers to study identifiers
trait_to_study(c('EFO_0005108', 'EFO_0005109'))

#> # A tibble: 2 x 2
#>   efo_id      study_id  
#>   <chr>       <chr>     
#> 1 EFO_0005108 GCST001762
#> 2 EFO_0005109 GCST001762

# Map EFO trait identifiers to association identifiers
trait_to_association(c('EFO_0005108', 'EFO_0005109'))

#> # A tibble: 38 x 2
#>    efo_id      association_id
#>    <chr>       <chr>         
#>  1 EFO_0005108 29052         
#>  2 EFO_0005108 29053         
#>  3 EFO_0005108 29054         
#>  4 EFO_0005108 29055         
#>  5 EFO_0005108 29165         
#>  6 EFO_0005108 29166         
#>  7 EFO_0005108 29167         
#>  8 EFO_0005108 29168         
#>  9 EFO_0005108 29169         
#> 10 EFO_0005109 29123         
#> # … with 28 more rows

# Map EFO trait identifiers to variant identifiers
trait_to_variant('EFO_0005229')

#> # A tibble: 30 x 2
#>    efo_id      variant_id
#>    <chr>       <chr>     
#>  1 EFO_0005229 rs4981419 
#>  2 EFO_0005229 rs16928927
#>  3 EFO_0005229 rs74702439
#>  4 EFO_0005229 rs35322295
#>  5 EFO_0005229 rs73481300
#>  6 EFO_0005229 rs1687482 
#>  7 EFO_0005229 rs6090818 
#>  8 EFO_0005229 rs73579354
#>  9 EFO_0005229 rs12881152
#> 10 EFO_0005229 rs182847  
#> # … with 20 more rows