Is the GWAS Catalog REST API server reachable?

Check if the EBI server where the GWAS Catalog REST API server is running is reachable. This function attempts to connect to https://www.ebi.ac.uk, returning TRUE on success, and FALSE otherwise. Set chatty = TRUE for a step by step description of the connection attempt.

Usage

is_ebi_reachable(url = "https://www.ebi.ac.uk", port = 443L, chatty = FALSE)

Arguments

url

NHGRI-EBI GWAS Catalog server URL. Default is https://www.ebi.ac.uk. You should not need to change this parameter.

port

Network port on which to ping the server. You should not need to change this parameter.

chatty

Whether to be verbose (TRUE) or not (FALSE).

Value

A logical value: TRUE if EBI server is reachable, FALSE otherwise.

Examples

# Check if the GWAS Catalog Server is reachable
is_ebi_reachable() # Returns TRUE or FALSE.
#> [1] TRUE

# Check if the GWAS Catalog Server is reachable
# and show exactly at what step is it failing (if that is the case)
is_ebi_reachable(chatty = TRUE)
#> Am I online? Yes.
#> Is there a DNS service running? Yes.
#> Connected to the internet? Yes.
#> Ping'ing www.ebi.ac.uk on port 443...
#> www.ebi.ac.uk successfully replied in 80.802 ms.
#> [1] TRUE