Number of GWAS Catalog entities

This function returns the number of unique entities in a GWAS Catalog object.

Usage

n(x, unique = FALSE)

# S4 method for class 'studies'
n(x, unique = FALSE)

# S4 method for class 'associations'
n(x, unique = FALSE)

# S4 method for class 'variants'
n(x, unique = FALSE)

# S4 method for class 'traits'
n(x, unique = FALSE)

Arguments

x

A studies, an associations, a variants, or a traits object.

unique

Whether to count only unique entries (TRUE) or not (FALSE).

Value

An integer scalar.

Examples

# Determine number of studies
n(studies_ex01)
#> [1] 2

# Determine number of associations
n(associations_ex01)
#> [1] 4

# Determine number of variants
n(variants_ex01)
#> [1] 3

# Determine number of traits
n(traits_ex01)
#> [1] 3