Creates a studies table.

Creates a studies table.

Usage

studies_tbl(
  study_id = character(),
  reported_trait = character(),
  initial_sample_size = character(),
  replication_sample_size = character(),
  gxe = logical(),
  gxg = logical(),
  snp_count = integer(),
  qualifier = character(),
  imputed = logical(),
  pooled = logical(),
  study_design_comment = character(),
  full_pvalue_set = logical(),
  user_requested = logical()
)

Arguments

study_id

GWAS Catalog study accession identifier.

reported_trait

Phenotypic trait as reported by the authors of the study.

initial_sample_size

Free text description of the initial cohort sample size.

replication_sample_size

Free text description of the replication cohort sample size.

gxe

Whether the study investigates a gene-environment interaction.

gxg

Whether the study investigates a gene-gene interaction.

snp_count

Number of variants passing quality control.

qualifier

Qualifier of number of variants passing quality control.

imputed

Whether variants were imputed.

pooled

Whether samples were pooled.

study_design_comment

Any other relevant study design information.

full_pvalue_set

Whether full summary statistics are available for this study.

user_requested

Whether the addition of this study to the GWAS Catalog was requested by a user.

Value

A tibble whose columns are the named arguments to the function.