Map a study id to a variant id

Map a study accession identifier to a variant accession identifier.

Usage

study_to_variant(study_id, verbose = FALSE, warnings = TRUE)

Arguments

study_id

A character vector of study accession identifiers.

verbose

Whether the function should be verbose about the different queries or not.

warnings

Whether to print warnings.

Value

A dataframe of two identifiers. First column is the study identifier and the second column is the variant identifier.

Examples

if (FALSE) { # \dontrun{
# Map GWAS study identifiers to variant identifiers
study_to_variant(c('GCST001084', 'GCST001085'))
} # }