Creates a variants table.

Creates a variants table.

Usage

variants_tbl(
  variant_id = character(),
  merged = integer(),
  functional_class = character(),
  chromosome_name = character(),
  chromosome_position = integer(),
  chromosome_region = character(),
  last_update_date = lubridate::ymd_hms()
)

Arguments

variant_id

A character vector of variant identifiers.

merged

A logical vector indicating if a SNP has been merged with another SNP in a newer genome build.

functional_class

A character vector of functional classes, see functional_class in slot variants of variants.

chromosome_name

A character vector of chromosome names.

chromosome_position

An integer vector of chromosome positions.

chromosome_region

A character vector of cytogenetic regions.

last_update_date

A POSIXct object indicating the last time the variants have been updated.

Value

A tibble whose columns are the named arguments to the function.