GRCh38 human cytogenetic bands.

A dataset containing the GRCh38 human cytogenetic bands and their genomic coordinates.

cytogenetic_bands

Format

A data frame with 862 rows and 7 variables:

cytogenetic_band

Cytogenetic band name. See Cytogenetic Nomenclature below.

chromosome

Chromosome name: 1 through 22 (the autosomes), X or Y.

start

Genomic start position of the cytogenetic band. Starts at 1.

end

Genomic end position of the cytogenetic band. End position is included in the band interval.

length

Length of the genomic interval of cytogenetic band.

assembly

Assembly version, should be 'GRCh38'.

stain

Giemsa stain results: Giemsa negative, 'gneg'; Giemsa positive, of increasing intensities, 'gpos25', 'gpos50', 'gpos75', and 'gpos100'; centromeric region, 'acen'; heterochomatin, either pericentric or telomeric, 'gvar'; and short arm of acrocentric chromosomes 13, 14, 15, 21, and 22 are coded as 'stalk'.

last_download_date

Time stamp of last time this dataset was downloaded from Ensembl.

Source

https://rest.ensembl.org/info/assembly/homo_sapiens?content-type=application/json&bands=1

Details

Genomic coordinates are for fully closed intervals.

Cytogenetic Nomenclature

Cytogenetic bands are numbered from the centromere outwards in both directions towards the telomeres on the shorter p arm and the longer q arm.

The first number or letter represents the chromosome. Chromosomes 1 through 22 (the autosomes) are designated by their chromosome number. The sex chromosomes are designated by X or Y. The next letter represents the arm of the chromosome: p or q.

The numbers cannot be read in the normal decimal numeric system e.g. 36, but rather 3-6 (region 3 band 6). Counting starts at the centromer as region 1 (or 1-0), to 11 (1-1) to 21 (2-1) to 22 (2-2) etc. Subbands are added in a similar way, eg. 21.1 to 21.2, if the bands are small or only appear at a higher resolution.