Request a paginated GWAS Catalog REST API endpoint

Performs a GET request on the specified resource_url and all its pages.

Usage

gc_request_all(
  resource_url = "/",
  base_url = gwas_rest_api_base_url,
  page_size = 20L,
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

resource_url

Endpoint URL. The endpoint is internally appended to the base_url. It should start with a forward slash (/).

base_url

The GWAS REST API base URL (one should not need to change its default value).

page_size

Page parameter used in the URL endpoint.

verbose

Whether to be chatty.

warnings

Whether to print warnings.

progress_bar

Whether to show a progress bar as the paginated resources are retrieved.

Value

A list four named elements:

url

The URL endpoint.

response_code

HTTP status code.

status

A string describing the status of the response obtained. It is "OK" if everything went OK or some other string describing the problem otherwise.

content

The parsed JSON as a nested list, as returned by jsonlite::fromJSON.