Get GWAS Catalog variants

Retrieves variants via the NHGRI-EBI GWAS Catalog REST API. The REST API is queried multiple times with the criteria passed as arguments (see below). By default all variants that match the criteria supplied in the arguments are retrieved: this corresponds to the default option set_operation set to 'union'. If you rather have only the variants that match simultaneously all criteria provided, then set set_operation to 'intersection'.

Usage

get_variants(
  study_id = NULL,
  association_id = NULL,
  variant_id = NULL,
  efo_id = NULL,
  pubmed_id = NULL,
  genomic_range = NULL,
  cytogenetic_band = NULL,
  gene_name = NULL,
  efo_trait = NULL,
  reported_trait = NULL,
  set_operation = "union",
  interactive = TRUE,
  std_chromosomes_only = TRUE,
  verbose = FALSE,
  warnings = TRUE
)

Arguments

study_id

A character vector of GWAS Catalog study accession identifiers.

association_id

A character vector of GWAS Catalog association identifiers.

variant_id

A character vector of GWAS Catalog variant identifiers.

efo_id

A character vector of EFO identifiers.

pubmed_id

An integer vector of PubMed identifiers.

genomic_range

A named list of three vectors:

chromosome

A character vector of chromosome names of the form 1--22, X or Y.

start

A numeric vector of start positions, starting at 1.

end

A numeric vector of end positions.

The three vectors need to be of the same length so that chromosome names, start and end positions can be matched by position.

cytogenetic_band

A character vector of cytogenetic bands of the form '1p36.11'.

gene_name

Gene symbol according to HUGO Gene Nomenclature (HGNC).

efo_trait

A character vector of EFO trait descriptions, e.g., 'uric acid measurement'.

reported_trait

A character vector of phenotypic traits as reported by the original authors of the study.

set_operation

Either 'union' or 'intersection'. This tells how variants retrieved by different criteria should be combined: 'union' binds together all results removing duplicates and 'intersection' only keeps same variants found with different criteria.

interactive

A logical. If all variants are requested, whether to ask interactively if we really want to proceed.

std_chromosomes_only

Whether to return only variants mapped to standard chromosomes: 1 thru 22, X, Y, and MT.

verbose

Whether the function should be verbose about the different queries or not.

warnings

Whether to print warnings.

Value

A variants object.

Details

Please note that all search criteria are vectorised, thus allowing for batch mode search, e.g., one can search by multiple variant identifiers at once by passing a vector of identifiers to variant_id.

Examples

# Get variants by study identifier
get_variants(study_id = 'GCST001085', warnings = FALSE)
#> An object of class "variants"
#> Slot "variants":
#> # A tibble: 18 × 7
#>    variant_id merged functional_cl…¹ chrom…² chrom…³ chrom…⁴ last_update_date   
#>    <chr>       <int> <chr>           <chr>     <int> <chr>   <dttm>             
#>  1 rs10496288      0 intergenic_var… NA      NA      NA      2021-06-24 19:41:17
#>  2 rs7827545       0 intron_variant  NA      NA      NA      2021-06-24 20:19:57
#>  3 rs10785581      0 intron_variant  12       4.55e7 12q12   2021-06-24 19:41:19
#>  4 rs200759        0 intron_variant  20       1.56e7 20p12.1 2021-06-24 19:41:19
#>  5 rs7960483       0 intron_variant  NA      NA      NA      2021-06-24 19:41:17
#>  6 rs1372662       0 intron_variant  8        1.35e8 8q24.22 2021-06-24 20:19:58
#>  7 rs10496289      0 intergenic_var… 2        8.31e7 2p12    2021-06-24 19:41:17
#>  8 rs6452524       0 intron_variant  NA      NA      NA      2021-06-24 19:41:17
#>  9 rs200752        0 intron_variant  NA      NA      NA      2021-06-24 19:41:17
#> 10 rs3798440       0 intron_variant  NA      NA      NA      2021-06-24 19:46:29
#> 11 rs7735940       0 regulatory_reg… NA      NA      NA      2021-06-24 19:41:11
#> 12 rs9350602       0 intron_variant  6        7.59e7 6q14.1  2021-06-24 19:46:30
#> 13 rs10188442      0 intron_variant  2        1.32e8 2q21.2  2021-06-24 20:19:59
#> 14 rs12522034      0 intergenic_var… 5        3.64e7 5p13.2  2021-06-24 19:41:12
#> 15 rs13420028      0 intron_variant  NA      NA      NA      2021-06-24 20:19:57
#> 16 rs2469997       0 intergenic_var… 8        1.19e8 8q24.12 2021-06-25 04:40:13
#> 17 rs6887846       0 intron_variant  5        8.31e7 5q14.2  2021-06-24 19:41:19
#> 18 rs6469823       1 intergenic_var… 8        1.19e8 8q24.12 2021-06-24 19:41:14
#> # … with abbreviated variable names ¹​functional_class, ²​chromosome_name,
#> #   ³​chromosome_position, ⁴​chromosome_region
#> 
#> Slot "genomic_contexts":
#> # A tibble: 60 × 12
#>    variant_id gene_name  chrom…¹ chrom…² dista…³ is_ma…⁴ is_cl…⁵ is_in…⁶ is_up…⁷
#>    <chr>      <chr>      <chr>     <int>   <int> <lgl>   <lgl>   <lgl>   <lgl>  
#>  1 rs10785581 LOC105369… 12       4.55e7   58636 FALSE   TRUE    TRUE    FALSE  
#>  2 rs10785581 ANO6       12       4.55e7   96786 FALSE   FALSE   TRUE    TRUE   
#>  3 rs10785581 ANO6       12       4.55e7   54910 TRUE    TRUE    TRUE    TRUE   
#>  4 rs10785581 U6         12       4.55e7   17400 TRUE    TRUE    TRUE    FALSE  
#>  5 rs10785581 LOC100128… 12       4.55e7   91755 FALSE   TRUE    TRUE    TRUE   
#>  6 rs10785581 LOC105369… 12       4.55e7       0 FALSE   FALSE   FALSE   FALSE  
#>  7 rs200759   MACROD2    20       1.56e7       0 FALSE   FALSE   FALSE   FALSE  
#>  8 rs200759   LOC107985… 20       1.56e7    1048 FALSE   TRUE    TRUE    FALSE  
#>  9 rs200759   ENSAP1     20       1.56e7    5991 FALSE   TRUE    TRUE    TRUE   
#> 10 rs200759   ENSAP1     20       1.56e7    5957 FALSE   TRUE    TRUE    TRUE   
#> # … with 50 more rows, 3 more variables: is_downstream <lgl>, source <chr>,
#> #   mapping_method <chr>, and abbreviated variable names ¹​chromosome_name,
#> #   ²​chromosome_position, ³​distance, ⁴​is_mapped_gene, ⁵​is_closest_gene,
#> #   ⁶​is_intergenic, ⁷​is_upstream
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> Slot "ensembl_ids":
#> # A tibble: 28 × 3
#>    variant_id gene_name ensembl_id     
#>    <chr>      <chr>     <chr>          
#>  1 rs10785581 ANO6      ENSG00000177119
#>  2 rs10785581 U6        ENSG00000277091
#>  3 rs200759   MACROD2   ENSG00000172264
#>  4 rs200759   ENSAP1    ENSG00000224274
#>  5 rs1372662  ZFAT-AS1  ENSG00000248492
#>  6 rs1372662  ZFAT      ENSG00000066827
#>  7 rs9350602  MYO6      ENSG00000196586
#>  8 rs9350602  RNU6-155P ENSG00000252156
#>  9 rs9350602  Y_RNA     ENSG00000252254
#> 10 rs9350602  RNA5SP209 ENSG00000223169
#> # … with 18 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "entrez_ids":
#> # A tibble: 32 × 3
#>    variant_id gene_name    entrez_id
#>    <chr>      <chr>        <chr>    
#>  1 rs10785581 LOC105369744 105369744
#>  2 rs10785581 ANO6         196527   
#>  3 rs10785581 LOC100128248 100128248
#>  4 rs10785581 LOC105369743 105369743
#>  5 rs200759   MACROD2      140733   
#>  6 rs200759   LOC107985394 107985394
#>  7 rs200759   ENSAP1       170511   
#>  8 rs1372662  ZFAT-AS1     594840   
#>  9 rs1372662  ZFAT         57623    
#> 10 rs9350602  MYO6         4646     
#> # … with 22 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 

# Get a variant by its identifier
if (FALSE) {
get_variants(variant_id = 'rs3798440', warnings = FALSE)
}