An S4 class to represent a set of GWAS Catalog studies

The studies object consists of eight slots, each a table (tibble), that combined form a relational database of a subset of GWAS Catalog studies. Each study is an observation (row) in the studies table --- main table. All tables have the column study_id as primary key.

Slots

studies

study_id

GWAS Catalog study accession identifier, e.g., "GCST002735".

reported_trait

Phenotypic trait as reported by the authors of the study, e.g. "Breast cancer".

initial_sample_size

Free text description of the initial cohort sample size.

replication_sample_size

Free text description of the replication cohort sample size.

gxe

Whether the study investigates a gene-environment interaction.

gxg

Whether the study investigates a gene-gene interaction.

snp_count

Number of variants passing quality control.

qualifier

Qualifier of number of variants passing quality control.

imputed

Whether variants were imputed.

pooled

Whether samples were pooled.

study_design_comment

Any other relevant study design information.

full_pvalue_set

Whether full summary statistics are available for this study.

user_requested

Whether the addition of this study to the GWAS Catalog was requested by a user.

genotyping_techs

A tibble listing genotyping technologies employed in each study. Columns:

study_id

GWAS Catalog study accession identifier.

genotyping_technology

Genotyping technology employed, e.g. "Exome genotyping array", "Exome-wide sequencing", "Genome-wide genotyping array", "Genome-wide sequencing", or "Targeted genotyping array".

platforms

A tibble listing platforms used per study.

study_id

GWAS Catalog study accession identifier.

manufacturer

Platform manufacturer, e.g., "Affymetrix", "Illumina", or "Perlegen".

ancestries

A tibble listing ancestry of samples used in each study.

study_id

GWAS Catalog study accession identifier.

ancestry_id

Ancestry identifier.

type

Stage of the ancestry sample: either 'initial' or 'replication'.

number_of_individuals

Number of individuals comprising this ancestry sample.

ancestral_groups

A tibble listing ancestral groups used in each ancestry.

study_id

GWAS Catalog study accession identifier.

ancestry_id

Ancestry identifier.

ancestral_group

Genetic ancentry groups present in the sample.

countries_of_origin

A tibble listing countries of origin of samples.

study_id

GWAS Catalog study accession identifier.

ancestry_id

Ancestry identifier.

country_name

Country name, according to The United Nations M49 Standard of Geographic Regions.

major_area

Region name, according to The United Nations M49 Standard of Geographic Regions.

region

Sub-region name, according to The United Nations M49 Standard of Geographic Regions.

countries_of_recruitment

A tibble listing countries of recruitment of samples.

study_id

GWAS Catalog study accession identifier.

ancestry_id

Ancestry identifier.

country_name

Country name, according to The United Nations M49 Standard of Geographic Regions.

major_area

Region name, according to The United Nations M49 Standard of Geographic Regions.

region

Sub-region name, according to The United Nations M49 Standard of Geographic Regions.

publications

A tibble listing publications associated with each study.

study_id

GWAS Catalog study accession identifier.

pubmed_id

PubMed identifier.

publication_date

Publication date (online date if available) formatted as ymd.

publication

Abbreviated journal name.

title

Publication title.

author_fullname

Last name and initials of first author.

author_orcid

Author's ORCID iD (Open Researcher and Contributor ID).