Subset a variants object

You can subset variants by identifier or by position using the `[` operator.

Usage

# S4 method for variants,missing,missing,missing
[(x, i, j, ..., drop = FALSE)

# S4 method for variants,numeric,missing,missing
[(x, i, j, ..., drop = FALSE)

# S4 method for variants,character,missing,missing
[(x, i, j, ..., drop = FALSE)

Arguments

x

A variants object.

i

Position of the identifier or the name of the identifier itself.

j

Not used.

...

Additional arguments not used here.

drop

Not used.

Value

A variants object.

Examples

# Subset a variants object by identifier
variants_ex01['rs4725504']
#> An object of class "variants"
#> Slot "variants":
#> # A tibble: 1 × 7
#>   variant_id merged functional_class chrom…¹ chrom…² chrom…³ last_update_date   
#>   <chr>       <int> <chr>            <chr>     <int> <chr>   <dttm>             
#> 1 rs4725504       0 intergenic_vari… 7        1.54e8 7q36.2  2019-05-20 22:10:00
#> # … with abbreviated variable names ¹​chromosome_name, ²​chromosome_position,
#> #   ³​chromosome_region
#> 
#> Slot "genomic_contexts":
#> # A tibble: 4 × 11
#>   varia…¹ gene_…² chrom…³ chrom…⁴ dista…⁵ is_cl…⁶ is_in…⁷ is_up…⁸ is_do…⁹ source
#>   <chr>   <chr>   <chr>     <int>   <int> <lgl>   <lgl>   <lgl>   <lgl>   <chr> 
#> 1 rs4725… DPP6    7        1.54e8  278759 TRUE    TRUE    FALSE   TRUE    Ensem…
#> 2 rs4725… DPP6    7        1.54e8  139795 TRUE    TRUE    FALSE   TRUE    NCBI  
#> 3 rs4725… PAXBP1… 7        1.54e8  158463 TRUE    TRUE    TRUE    FALSE   Ensem…
#> 4 rs4725… LOC107… 7        1.54e8   84107 TRUE    TRUE    TRUE    FALSE   NCBI  
#> # … with 1 more variable: mapping_method <chr>, and abbreviated variable names
#> #   ¹​variant_id, ²​gene_name, ³​chromosome_name, ⁴​chromosome_position, ⁵​distance,
#> #   ⁶​is_closest_gene, ⁷​is_intergenic, ⁸​is_upstream, ⁹​is_downstream
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "ensembl_ids":
#> # A tibble: 2 × 3
#>   variant_id gene_name ensembl_id     
#>   <chr>      <chr>     <chr>          
#> 1 rs4725504  DPP6      ENSG00000130226
#> 2 rs4725504  PAXBP1P1  ENSG00000233489
#> 
#> Slot "entrez_ids":
#> # A tibble: 3 × 3
#>   variant_id gene_name    entrez_id
#>   <chr>      <chr>        <chr>    
#> 1 rs4725504  DPP6         1804     
#> 2 rs4725504  PAXBP1P1     106481704
#> 3 rs4725504  LOC107986750 107986750
#> 

# Or by its position in table variants
variants_ex01[3]
#> An object of class "variants"
#> Slot "variants":
#> # A tibble: 1 × 7
#>   variant_id merged functional_class chrom…¹ chrom…² chrom…³ last_update_date   
#>   <chr>       <int> <chr>            <chr>     <int> <chr>   <dttm>             
#> 1 rs4725504       0 intergenic_vari… 7        1.54e8 7q36.2  2019-05-20 22:10:00
#> # … with abbreviated variable names ¹​chromosome_name, ²​chromosome_position,
#> #   ³​chromosome_region
#> 
#> Slot "genomic_contexts":
#> # A tibble: 4 × 11
#>   varia…¹ gene_…² chrom…³ chrom…⁴ dista…⁵ is_cl…⁶ is_in…⁷ is_up…⁸ is_do…⁹ source
#>   <chr>   <chr>   <chr>     <int>   <int> <lgl>   <lgl>   <lgl>   <lgl>   <chr> 
#> 1 rs4725… DPP6    7        1.54e8  278759 TRUE    TRUE    FALSE   TRUE    Ensem…
#> 2 rs4725… DPP6    7        1.54e8  139795 TRUE    TRUE    FALSE   TRUE    NCBI  
#> 3 rs4725… PAXBP1… 7        1.54e8  158463 TRUE    TRUE    TRUE    FALSE   Ensem…
#> 4 rs4725… LOC107… 7        1.54e8   84107 TRUE    TRUE    TRUE    FALSE   NCBI  
#> # … with 1 more variable: mapping_method <chr>, and abbreviated variable names
#> #   ¹​variant_id, ²​gene_name, ³​chromosome_name, ⁴​chromosome_position, ⁵​distance,
#> #   ⁶​is_closest_gene, ⁷​is_intergenic, ⁸​is_upstream, ⁹​is_downstream
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "ensembl_ids":
#> # A tibble: 2 × 3
#>   variant_id gene_name ensembl_id     
#>   <chr>      <chr>     <chr>          
#> 1 rs4725504  DPP6      ENSG00000130226
#> 2 rs4725504  PAXBP1P1  ENSG00000233489
#> 
#> Slot "entrez_ids":
#> # A tibble: 3 × 3
#>   variant_id gene_name    entrez_id
#>   <chr>      <chr>        <chr>    
#> 1 rs4725504  DPP6         1804     
#> 2 rs4725504  PAXBP1P1     106481704
#> 3 rs4725504  LOC107986750 107986750
#> 

# Keep all variants except the third
variants_ex01[-3]
#> An object of class "variants"
#> Slot "variants":
#> # A tibble: 2 × 7
#>   variant_id  merged functional_cl…¹ chrom…² chrom…³ chrom…⁴ last_update_date   
#>   <chr>        <int> <chr>           <chr>     <int> <chr>   <dttm>             
#> 1 rs146992477      0 regulatory_reg… 8        3.70e7 8p11.23 2019-05-21 21:33:39
#> 2 rs56261590       0 intron_variant  4        7.32e7 4q13.3  2019-05-20 22:09:53
#> # … with abbreviated variable names ¹​functional_class, ²​chromosome_name,
#> #   ³​chromosome_position, ⁴​chromosome_region
#> 
#> Slot "genomic_contexts":
#> # A tibble: 17 × 11
#>    variant_id  gene_name chrom…¹ chrom…² dista…³ is_cl…⁴ is_in…⁵ is_up…⁶ is_do…⁷
#>    <chr>       <chr>     <chr>     <int>   <int> <lgl>   <lgl>   <lgl>   <lgl>  
#>  1 rs146992477 LOC10537… 8        3.70e7   81555 FALSE   TRUE    FALSE   TRUE   
#>  2 rs146992477 KCNU1     8        3.70e7   48413 TRUE    TRUE    TRUE    FALSE  
#>  3 rs146992477 LOC10537… 8        3.70e7    2533 TRUE    TRUE    TRUE    FALSE  
#>  4 rs146992477 TPT1P8    8        3.70e7   95231 FALSE   TRUE    TRUE    FALSE  
#>  5 rs146992477 RPL26P25  8        3.70e7    3392 TRUE    TRUE    FALSE   TRUE   
#>  6 rs146992477 KCNU1     8        3.70e7   48413 FALSE   TRUE    TRUE    FALSE  
#>  7 rs146992477 AC090453… 8        3.70e7    3439 TRUE    TRUE    FALSE   TRUE   
#>  8 rs146992477 TPT1P8    8        3.70e7   95008 FALSE   TRUE    TRUE    FALSE  
#>  9 rs146992477 AC092818… 8        3.70e7   82903 FALSE   TRUE    FALSE   TRUE   
#> 10 rs146992477 LOC10537… 8        3.70e7   59331 FALSE   TRUE    TRUE    FALSE  
#> 11 rs56261590  ANKRD17   4        7.32e7       0 FALSE   FALSE   FALSE   FALSE  
#> 12 rs56261590  ANKRD17   4        7.32e7       0 FALSE   FALSE   FALSE   FALSE  
#> 13 rs56261590  HMGA1P2   4        7.32e7   90636 TRUE    TRUE    TRUE    FALSE  
#> 14 rs56261590  HMGA1P2   4        7.32e7   89275 TRUE    TRUE    TRUE    FALSE  
#> 15 rs56261590  LOC10272… 4        7.32e7   69394 TRUE    TRUE    FALSE   TRUE   
#> 16 rs56261590  AC053527… 4        7.32e7   69428 TRUE    TRUE    FALSE   TRUE   
#> 17 rs56261590  RF00334   4        7.32e7   74179 FALSE   TRUE    FALSE   TRUE   
#> # … with 2 more variables: source <chr>, mapping_method <chr>, and abbreviated
#> #   variable names ¹​chromosome_name, ²​chromosome_position, ³​distance,
#> #   ⁴​is_closest_gene, ⁵​is_intergenic, ⁶​is_upstream, ⁷​is_downstream
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "ensembl_ids":
#> # A tibble: 8 × 3
#>   variant_id  gene_name  ensembl_id     
#>   <chr>       <chr>      <chr>          
#> 1 rs146992477 KCNU1      ENSG00000215262
#> 2 rs146992477 TPT1P8     ENSG00000219491
#> 3 rs146992477 AC090453.1 ENSG00000243503
#> 4 rs146992477 AC092818.1 ENSG00000254038
#> 5 rs56261590  ANKRD17    ENSG00000132466
#> 6 rs56261590  HMGA1P2    ENSG00000248641
#> 7 rs56261590  AC053527.1 ENSG00000250220
#> 8 rs56261590  RF00334    ENSG00000221093
#> 
#> Slot "entrez_ids":
#> # A tibble: 9 × 3
#>   variant_id  gene_name    entrez_id
#>   <chr>       <chr>        <chr>    
#> 1 rs146992477 LOC105379377 105379377
#> 2 rs146992477 KCNU1        157855   
#> 3 rs146992477 LOC105379376 105379376
#> 4 rs146992477 TPT1P8       59347    
#> 5 rs146992477 RPL26P25     100270966
#> 6 rs146992477 LOC105379375 105379375
#> 7 rs56261590  ANKRD17      26057    
#> 8 rs56261590  HMGA1P2      171559   
#> 9 rs56261590  LOC102724832 102724832
#>