An S4 class to represent a set of GWAS Catalog variants

The variants object consists of four slots, each a table (tibble), that combined form a relational database of a subset of GWAS Catalog variants Each variant is an observation (row) in the variants table --- main table. All tables have the column variant_id as primary key.

Slots

variants

A tibble listing variants. Columns:

variant_id

Variant identifier, e.g., 'rs1333048'.

merged

Whether this SNP has been merged with another SNP in a newer genome build.

functional_class

Class according to Ensembl's predicted consequences that each variant allele may have on transcripts. See Ensembl Variation - Calculated variant consequences.

chromosome_name

Chromosome name.

chromosome_position

Chromosome position.

chromosome_region

Cytogenetic location.

last_update_date

Last time this variant was updated.

genomic_contexts

A tibble listing genomic contexts associated with each variant. Columns:

variant_id

Variant identifier.

gene_name

Gene symbol according to HUGO Gene Nomenclature (HGNC).

chromosome_name

Chromosome name.

chromosome_position

Chromosome position.

distance

Genomic distance between the variant and the gene (in base pairs).

is_closest_gene

Whether this is the closest gene to this variant.

is_intergenic

Whether this variant is intergenic, i.e, if there is no gene up or downstream within 100kb.

is_upstream

Whether this variant is upstream of this gene.

is_downstream

Whether this variant is downstream of this gene.

source

Gene mapping source, either Ensembl or NCBI.

mapping_method

Gene mapping method.

ensembl_ids

A tibble listing gene Ensembl identifiers associated with each genomic context. Columns:

variant_id

Variant identifier.

gene_name

Gene symbol according to HUGO Gene Nomenclature (HGNC).

ensembl_id

The Ensembl identifier of an Ensembl gene, see Section Gene annotation in Ensembl for more information.

entrez_ids

A tibble listing gene Entrez identifiers associated with each genomic context. Columns:

variant_id

Variant identifier.

gene_name

Gene symbol according to HUGO Gene Nomenclature (HGNC).

entrez_id

The Entrez identifier of a gene, see ref. 10.1093/nar/gkq1237 for more information.